requirements.txt
# Core dependencies
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM providers
google-generativeai>=0.3.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.1.0

# Vector database
# CRITICAL: Pinned to v3.x because Weaviate Server v1.24.1 is incompatible with v4 client
weaviate-client>=3.26.7,<4.0.0

# Embeddings and reranking
sentence-transformers>=3.0.0
# Note: For GPU support, usually installed via --index-url, but this sets the version baseline
torch>=2.5.1
torchvision>=0.20.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
pylint>=2.17.0
mypy>=1.4.0

# Documentation
sphinx>=7.0.0